fixed handling of progress, now as double instead of long

+++ src/main/java/edu/brown/cs/student/weekli/main/Handlers.java
@@ -156,17 +156,17 @@ public class Handlers {
             User loggedIn = (User) session.getAttribute("user");
             if (id.equals("") && progress.equals("")) {
                 List<Task> c = loggedIn.updateSchedule();
-                complete = c.stream().map(t -> new String[]{t.getName(), t.getDescription(), t.getID().toString(), Long.toString(t.getProgress()), t.getColor()}).collect(Collectors.toList());
-                tasks = loggedIn.getTasks().stream().map(t -> new String[]{t.getName(), t.getDescription(), t.getID().toString(), Long.toString(t.getProgress()), t.getColor()}).collect(Collectors.toList());
+                complete = c.stream().map(t -> new String[]{t.getName(), t.getDescription(), t.getID().toString(), Double.toString(t.getProgress()), t.getColor()}).collect(Collectors.toList());
+                tasks = loggedIn.getTasks().stream().map(t -> new String[]{t.getName(), t.getDescription(), t.getID().toString(), Double.toString(t.getProgress()), t.getColor()}).collect(Collectors.toList());
                 // send back both completed tasks as well as all tasks
                 variables = ImmutableMap.of("complete", complete, "tasks", tasks);
             } else {
                 UUID ID = UUID.fromString(id);
-                long prog = Long.parseLong(progress);
+                double prog = Double.parseDouble(progress);
                 Task toUpdate = loggedIn.belongsToTask(ID);
                 toUpdate.setProgress(prog);
                 loggedIn.updateTaskInDB(toUpdate);
-                String[] task = new String[]{toUpdate.getName(), toUpdate.getDescription(), toUpdate.getID().toString(), Long.toString(toUpdate.getProgress()), toUpdate.getColor()};
+                String[] task = new String[]{toUpdate.getName(), toUpdate.getDescription(), toUpdate.getID().toString(), Double.toString(toUpdate.getProgress()), toUpdate.getColor()};
                 // send back updated task
                 variables = ImmutableMap.of("updated", task);
             }
